Added a gate delay parameter.
This parameter is set for each individual CV output and controls how many ticks the GATE_ON message is delayed by. In particular, this fixes issues when connecting CARBON to the Moog DFAM (CV -> velocity, GATE -> trigger). It seems as though the DFAM is either very intolerant to timing discrepancies between the velocity + trigger inputs, or that the DFAM itself is delayed in recognizing voltage changes to the velocity input. In either case, delaying the CARBON GATE ON message by about 3-4 ticks fixes the problem.

This parameter is saved with the song data and thus required an increment to the CARBON version (1.15 now). I've only done basic testing with a simple track using ABCD cv pairing. It would be interesting to see how this holds up in a polyphonic scenario, or even LIVE mode (which I haven't even used yet).